server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
fpdf2
